split up into blueprints and functions
